+++ src/main/java/gov/usgs/cida/wqp/transform/MapToKmlTransformer.java
@@ -3,7 +3,6 @@ package gov.usgs.cida.wqp.transform;
 import static gov.usgs.cida.wqp.mapping.xml.StationKml.KML_DOCUMENT;
 
 import java.io.OutputStream;
-import java.math.BigDecimal;
 
 import gov.usgs.cida.wqp.mapping.Profile;
 import gov.usgs.cida.wqp.mapping.xml.IXmlMapping;
@@ -11,7 +10,7 @@ import gov.usgs.cida.wqp.service.ILogService;
 
 public class MapToKmlTransformer extends MapToXmlTransformer {
 
-	public MapToKmlTransformer(OutputStream target, IXmlMapping mapping, ILogService logService, BigDecimal logId, Profile profile) {
+	public MapToKmlTransformer(OutputStream target, IXmlMapping mapping, ILogService logService, Integer logId, Profile profile) {
 		super(target, mapping, logService, logId, profile);
 	}
